Side Menu Upgrade - Audio Reactive 1/2
Part one of two, trigger the effects in the side menu by pressing `q` on your keyboard!

Not audio reactive at the moment, but it should be pretty straightforward for me to do FFT analysis on any active audio sources on the website, and color-match along with determining how much splats to be added.